Prints out the timestamp

Prints out the timestamp to console and to a file. Useful for diagnosing potential failures of a terminal server session.

Usage

count_up_till_death(
  file_out,
  extrainfo = NULL,
  secs_interval = 60,
  max_lines = 10000,
  verbose = TRUE
)

Arguments

file_out

Character, path to the filename to print out the series of timestamps

extrainfo

Character, specifies the extra information to print out before the timestamp is recorded. Defaults to

secs_interval

Numeric value, specifying the interval of seconds after which to repeat the timestamp operation

max_lines

Integer, number of max lines to record - the time interval covered will depend on this and on the secs_interval. This is mainly to avoid having a really endless loop and not being able to stop the process.

verbose

Logical, defaults to TRUE

Value

Invisible NULL - the output is very much the side effect of printing the timestamp to file

Examples

# count_up_till_death("mytestfile.txt", extrainfo = "YourName")