Title

Title

Usage

create_DEresults(
  resuSet = NULL,
  dds_obj,
  contrast_name,
  FDR = 0.05,
  lfc_threshold = 0,
  extended_anno_df,
  species = "Homo_sapiens",
  name_res_entry = contrast_name,
  verbose = TRUE
)

Arguments

resuSet

A list, provided in the format expected by the resuSet specifications. This can be an empty list, if the object was never created beforehand, otherwise this function can simply add all the set of the results and output into a new element of an existing list. Defaults to NULL, which would lead to the creation of a new list object. If this is not the desired behavior, please provide an existing object structured as a resuSet.

dds_obj

A DESeqDataset object, meaningfully related/matched to the resuSet object.

contrast_name

Character string, specifying the name of the contrast to be extracted

FDR

Numeric value, specifying the false discovery rate value to be used

lfc_threshold

Numeric value, corresponding to the logFC theshold to be used to extract the results - it is relevant in the call to results() and it does differ (a lot, conceptually) from the logFC filtering that commonly is performed post-hoc, after testing against the default null hypothesis (logFC = 0), which might be too liberal in some cases.

extended_anno_df

An annotation data.frame with the following columns: gene_id and gene_name (mandatory), and description (optional, can be retrieved via biomaRt)

species

Character string, specifying the species for the dataset under investigation. Defaults to "Homo_sapiens", could also be "Mus_musculus" and other similarly assembled character strings. This is relevant only when creating the buttons for the interactive versions of the results tables.

name_res_entry

Character string

verbose

Logical

Value

A resuSet list object.

Examples

# TODO