Fortify different set of annotations

Fortify different set of annotations

Usage

fortify_annotations(
  anno_df_1,
  anno_df_2,
  id_1 = "gene_id",
  id_2 = "ensembl_gene_id",
  gene_name_1 = "gene_name",
  gene_name_2 = "external_gene_name",
  dds = NULL,
  verbose = TRUE
)

Arguments

anno_df_1

A data.frame object, corresponding to the first annotation table

anno_df_2

A data.frame object, corresponding to the second annotation table

id_1

Character string, specifying which column in the first table is to be used as identifier (defaults to gene_id)

id_2

Character string, specifying which column in the second table is to be used as identifier (defaults to ensembl_gene_id, for biomaRt-like objects)

gene_name_1

Character string, specifying which column in the first table is to be used as gene name (defaults to gene_name)

gene_name_2

Character string, specifying which column in the second table is to be used as identifier (defaults to external_gene_name, again as could be taken out e.g. from biomaRt)

dds

A DESeqDataset object

verbose

Logical, controlling the verbosity level of the function

Value

A list, including the fortified annotation data frame (in the anno_df element), and (if provided as input) a dds object (in the dds object) with the information stored in the rowData slot.

Examples

# TODO