M&Ms maker

Helping you to draft the materials and methods section

Usage

matmetmaker_RNAseqbulk(
  dds_object,
  template_text = NULL,
  packages = NULL,
  FDRthreshold = NULL,
  de_framework = "DESeq2",
  add_citations = TRUE
)

Arguments

dds_object

A DESeqDataset object. Necessary to read in the info on how the quantifications were performed. Thanks tximeta!

template_text

A (long, correctly structured) character string. It defaults to NULL, which would then lead to using the "default" content. TODO: To implement and possibly use "custom"/personalized ones, that could ideally live in separated files (one for each? one for type?)

packages

Character vector. Defaults to NULL. Specifies which packages would need to be captured, alongside the "standard" ones.

FDRthreshold

Numeric value, specifying a precise FDR value. If left unspecified (NULL by default) it searches for the value in the environment of the variable called FDR.

de_framework

Character string, defaults to "DESeq2"- if implemented in other frameworks we could use this to swap some behavior.

add_citations

Logical value, TODO to be implemented. In its implementation it should add after the tools the relevant DOIs and/or PMIDs to the articles it should refer to. This should simplify our lives and the lives of our collaborators to add the relevant references into the manuscripts.

Value

A character fairly long string, which can be used to draft quickly the M&M section

Examples

# TODO