Title

Title

Usage

samplesheet_creator(
  n_samples = 4,
  ss_filename,
  ss_dir = ".",
  extra_fields = NULL
)

Arguments

n_samples

Numeric value, corresponds to the number of samples included

ss_filename

Character string, basename of the excel file that will get generated

ss_dir

Character string, folder where the sample sheet will be generated into. Defaults to "." (the current working directory).

extra_fields

Vector of character strings, specifying the names of the columns for extra fields to be added.

Value

The character string corresponding to the path to the Excel sample sheet created. The Excel sheet is created as a side effect.

Examples

mysheet <- samplesheet_creator(ss_filename = "mysheet",
                               ss_dir = tempdir())
#> Created a new sample sheet for 4 samples with the following fields:
#> id, fastq_file1, fastq_file2, group, project_scenario, sample_id, quants_files, notes
#> ------
#> Please make sure to fill in the respective info fields, 
#> following the indications in the pre-filled cells
mysheet
#> [1] "/private/var/folders/sx/mf6nj_sn1ll2crpml02qjgbr0000gn/T/RtmpBcNfef/mysheet.xlsx"